+++ packages/most_key_value_editor/lib/src/utils/most_property_ext.dart
@@ -1,6 +1,7 @@
 import 'package:most_schema_parser/most_schema_parser.dart';
 import 'package:recase/recase.dart';
 
+/// Display title extension
 extension MostPropertyExtension on MostProperty {
   /// Display title that is visible to user.
   String get displayTitle => title ?? propertyName.titleCase;