+++ packages/most_key_value_editor/CHANGELOG.md
@@ -1,3 +1,13 @@
+## 1.0.0
+- BREAKING CHANGE: `JsonReaderWriter` renamed to `JsonAccessor`.
+- BREAKING CHANGE: `JsonPropertyReaderWriter` renamed to `JsonPropertyAccessor`.
+- BREAKING CHANGE: `JsonAccessor` is not holding command history (i.e. undo feature is temporary removed).
+- BREAKING CHANGE: `MostKeyValueEditor#customBuilders` replaced with 3 dedicated parameters.
+- BREAKING CHANGE: `MostKeyValueEditorController` is not supporting validation using FormKey.
+- BREAKING CHANGE: `MostKeyValueEditorController` validates JSON when it is changed.
+- BREAKING CHANGE: `StringFormField` renamed to `StringField` and saves value only on user interaction (i.e. tap).
+- BREAKING CHANGE: `MostValidator` renamed to `EditorStateValidator`.
+
 ## 0.0.2
 - Add `MostValidator` concept:
   - Allows to add custom validators to the key-value editor.